completed elevator refactor
