+++ src/main/java/frc/robot/commands/CompositeCommands.java
@@ -11,9 +11,9 @@ import frc.robot.RobotStateLL;
 import frc.robot.subsystems.shared.climber.Climber;
 import frc.robot.subsystems.shared.climber.ClimberConstants;
 import frc.robot.subsystems.shared.drive.Drive;
-import frc.robot.subsystems.shared.elevator.ElevatorCSB;
+import frc.robot.subsystems.shared.elevator.Elevator.ElevatorCSB;
+import frc.robot.subsystems.shared.elevator.Elevator.ElevatorFSM;
 import frc.robot.subsystems.shared.elevator.ElevatorConstants.ElevatorPositions;
-import frc.robot.subsystems.shared.elevator.ElevatorFSM;
 import frc.robot.subsystems.shared.funnel.FunnelCSB;
 import frc.robot.subsystems.shared.funnel.FunnelConstants.FunnelState;
 import frc.robot.subsystems.shared.visionlimelight.Camera;
