+++ src/main/java/frc/robot/util/LTNUpdater.java
@@ -3,9 +3,9 @@ package frc.robot.util;
 import frc.robot.commands.DriveCommands;
 import frc.robot.subsystems.shared.drive.Drive;
 import frc.robot.subsystems.shared.drive.DriveConstants;
-import frc.robot.subsystems.shared.elevator.ElevatorCSB;
+import frc.robot.subsystems.shared.elevator.Elevator.ElevatorCSB;
+import frc.robot.subsystems.shared.elevator.Elevator.ElevatorFSM;
 import frc.robot.subsystems.shared.elevator.ElevatorConstants;
-import frc.robot.subsystems.shared.elevator.ElevatorFSM;
 import frc.robot.subsystems.shared.funnel.FunnelCSB;
 import frc.robot.subsystems.shared.funnel.FunnelConstants;
 import frc.robot.subsystems.shared.funnel.FunnelFSM;
